Refactor move collection extensions to kstore
